requirements.txt
langchain
langchain-community
langchain-groq
# groq
langchain-weaviate

langchainhub
weaviate-client>=4.0.0
sentence-transformers
scikit-learn
streamlit
mlflow
huggingface-hub